cleaner use of getattribute
