+++ tests/test_fields.py
@@ -293,72 +293,6 @@ class TestCreateOnlyDefault:
         }
 
 
-# Tests for RecursiveField.
-# -------------------------
-
-class TestRecursiveField:
-    def setup(self):
-        class LinkSerializer(serializers.Serializer):
-            name = serializers.CharField()
-            next = serializers.RecursiveField(required=False, allow_null=True)
-        self.link_serializer = LinkSerializer
-
-        class NodeSerializer(serializers.Serializer):
-            name = serializers.CharField()
-            children = serializers.ListField(child=serializers.RecursiveField())
-        self.node_serializer = NodeSerializer
-
-    def test_link_serializer(self):
-        value = {
-            'name': 'first',
-            'next': {
-                'name': 'second',
-                'next': {
-                    'name': 'third',
-                    'next': None,
-                }
-            }
-        }
-
-        # test serialization
-        serializer = self.link_serializer(value)
-
-        assert serializer.data == value, \
-            'serialized data does not match input'
-
-        # test deserialization
-        serializer = self.link_serializer(data=value)
-
-        assert serializer.is_valid(), \
-            'cannot validate on deserialization: %s' % dict(serializer.errors)
-        assert serializer.validated_data == value, \
-            'deserialized data does not match input'
-
-    def test_node_serializer(self):
-        value = {
-            'name': 'root',
-            'children': [{
-                'name': 'first child',
-                'children': [],
-            }, {
-                'name': 'second child',
-                'children': [],
-            }]
-        }
-
-        # serialization
-        serializer = self.node_serializer(value)
-        assert serializer.data == value, \
-            'serialized data does not match input'
-
-        # deserialization
-        serializer = self.node_serializer(data=value)
-        assert serializer.is_valid(), \
-            'cannot validate on deserialization: %s' % dict(serializer.errors)
-        assert serializer.validated_data == value, \
-            'deserialized data does not match input'
-
-
 # Tests for field input and output values.
 # ----------------------------------------
